feature: рефакторингдобавÐление компонента вложенной нÐтаблицы AlbumPreview
